Added concept of pure virtual functions. More porting.
Pure virtual functions are notated with @virtual. The prototype should be notated as @abstract. The value of the function can be set to PlurObject.pureVirtualFunction to help enforce implementation. If so assigned, a call to an unimplemented pure virtual function will throw an error. If PLUR_DEBUG is true, then prototype creation will add a check for unimplemented virtual functions and throw an error as well.

Moved message to msg and simplified the naming scheme there.

msg/Message created as an ABC that both msg/Request and msg/Response are extended from.

JqueryRequestClient moved to new plur/jquery heirarchy.

Handful of other standardization updates.